Implemented a function that gets newly inserted data
This resolves #23, running the test requires nivacloud to be up and
running. The function 'get_newly_inserted_data' requires a start and an
end-time. It can return both aggregated and unaggregated data. To work
the function requires both the metaflow and tsb url.

The requests involved to both endpoints use the same traceid and header.