Added a new test that communicates via ms-ferrybox
The new test requires a valid token to communicate with the cloud api.
It is quite slow and takes about 5 minutes to complete in it's current
form. To speed this up there needs to be an index in TSB and we need to
query a timerange that doesn't include the moment the timestamp column
was created. The timestamp '2020-06-15 10:56:39.324063' includes almost
all data right now.